Refactored queryNode function
I simplified the queryNode function code and changed its signature a
bit. Now the user can query for all marginals or just a set of nodes of
interest. As a consequence, the hasNode function was removed. Moreover,
the queryNodeExample.R was also changed in order to reflect the changes
made to the corresponding function.

I also changed the name of the function parentsDefined to
isParentsDefine to reflect the fact that it returns a boolean.